+++ avs/app/src/main/java/org/example/age/avs/app/AvsApp.java
@@ -10,11 +10,11 @@ import org.example.age.avs.api.AvsApi;
 import org.example.age.avs.app.config.AvsAppConfig;
 import org.example.age.avs.app.config.AvsConfigModule;
 import org.example.age.avs.client.site.SiteClientsModule;
+import org.example.age.avs.provider.accountstore.dynamodb.DynamoDbAvsAccountStoreModule;
 import org.example.age.common.app.env.DropwizardEnvModule;
 import org.example.age.common.provider.account.demo.DemoAccountIdModule;
 import org.example.age.common.provider.pendingstore.redis.RedisPendingStoreModule;
 import org.example.age.module.crypto.demo.DemoAvsCryptoModule;
-import org.example.age.module.store.dynamodb.DynamoDbAvsAccountStoreModule;
 import org.example.age.service.AvsServiceModule;
 
 /** Application for the age verification service. */
