+++ site/app/src/main/java/org/example/age/site/app/SiteApp.java
@@ -10,12 +10,12 @@ import org.example.age.common.app.env.DropwizardEnvModule;
 import org.example.age.common.provider.account.demo.DemoAccountIdModule;
 import org.example.age.common.provider.pendingstore.redis.RedisPendingStoreModule;
 import org.example.age.module.crypto.demo.DemoSiteCryptoModule;
-import org.example.age.module.store.dynamodb.DynamoDbSiteAccountStoreModule;
 import org.example.age.service.SiteServiceModule;
 import org.example.age.site.api.SiteApi;
 import org.example.age.site.app.config.SiteAppConfig;
 import org.example.age.site.app.config.SiteConfigModule;
 import org.example.age.site.client.avs.AvsClientModule;
+import org.example.age.site.provider.accountstore.dynamodb.DynamoDbSiteAccountStoreModule;
 
 /** Application for a site. */
 public final class SiteApp extends Application<SiteAppConfig> {
