+++ site/provider/accountstore-dynamodb/src/test/java/org/example/age/site/provider/accountstore/dynamodb/DynamoDbSiteAccountStoreTest.java
@@ -1,0 +1,44 @@
+package org.example.age.site.provider.accountstore.dynamodb;
+
+import dagger.BindsInstance;
+import dagger.Component;
+import jakarta.inject.Singleton;
+import java.util.function.Supplier;
+import org.example.age.common.client.dynamodb.DynamoDbClientConfig;
+import org.example.age.site.spi.SiteVerificationStore;
+import org.example.age.testing.client.TestClient;
+import org.example.age.testing.env.TestEnvModule;
+import org.example.age.testing.site.spi.SiteAccountStoreTestTemplate;
+import software.amazon.awssdk.regions.Region;
+
+public final class DynamoDbSiteAccountStoreTest extends SiteAccountStoreTestTemplate {
+
+    private static final SiteVerificationStore store = TestComponent.create();
+
+    @Override
+    protected SiteVerificationStore store() {
+        return store;
+    }
+
+    /** Dagger component for {@link SiteVerificationStore} */
+    @Component(modules = {DynamoDbSiteAccountStoreModule.class, TestEnvModule.class})
+    @Singleton
+    interface TestComponent extends Supplier<SiteVerificationStore> {
+
+        static SiteVerificationStore create() {
+            DynamoDbClientConfig config = DynamoDbClientConfig.builder()
+                    .region(Region.US_EAST_1.toString())
+                    .testEndpointOverride(TestClient.localhostUri(8000))
+                    .build();
+            return DaggerDynamoDbSiteAccountStoreTest_TestComponent.factory()
+                    .create(config)
+                    .get();
+        }
+
+        @Component.Factory
+        interface Factory {
+
+            TestComponent create(@BindsInstance DynamoDbClientConfig config);
+        }
+    }
+}
